feat: allow different converters than Avro
